+++ lib/hir-mir/src/variant_type_collection.rs
